Add CodeChecker-like workflows
We discussed that maybe the tests relating to these rules have a better
place here than in the CodeChecker repository itself. I mostly
copy-pasted the relevant files from the .github/ folder from CodeChecker
to this repo, and migrated the tests to pytest. Not sure if I aced any
logging info of value, that is worth the most attention in this PR.